New command: parabody.
This is a hack to make the font size that is used in \location bodies
available without the parameter.

The font size that is used by location becomes the normal font affecting
the body of other parts (like \section).  Sections that are placed
before a \location is used have a different font size.

I don't yet fully understand the application of the font sizes, so for
now I'm just creating a new component.